CDS Hooks(プロトタイプ版) の活用事例/LLMService/requirements.txt
Flask
flask-cors
python-dotenv
cohere
openai
google-generativeai
